# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Finance API
  description: |
    API for the Personal Finance application. Provides endpoints for managing
    bank accounts, transactions, investments, market data, and notifications.

    ## Authentication
    Most endpoints require authentication via Supabase Auth. Include the session
    cookie or Authorization header with a valid JWT token.

    ## Response Format
    Successful responses follow this format:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "count": 10 }
    }
    ```

    Error responses follow this format:
    ```json
    {
      "success": false,
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: API routes
...
paths:
...
  /market/analysis:
    get:
      tags: [Market]
      summary: Get market analysis
      description: Returns AI-generated market analysis
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Market analysis
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /investments/holdings:
    get:
      tags: [Investments]
      summary: Get investment holdings
      description: Returns detailed investment holdings
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Investment holdings
          content:
            application/json:
              schema:
                type: object
                properties:
                  holdings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Holding'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ DISCORD ============
  /discord/register:
    post:
      tags: [Discord]
      summary: Register Discord user
      description: Links a Discord user to their finance account
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                discordUserId:
                  type: string
                guildId:
                  type: string
      responses:
        '200':
          description: Registration successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /discord/guilds:
    get:
      tags: [Discord]
      summary: List Discord guilds
      description: Returns Discord guilds the bot is in
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of guilds
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for scheduled jobs
    discordSignature:
      type: apiKey
      in: header
      name: X-Signature-Ed25519
      description: Discord interaction signature
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Unauthorized"
...
  schemas:
...
    Holding:
      type: object
      properties:
        symbol:
          type: string
        shares:
          type: number
        costBasis:
          type: number
        currentValue:
          type: number
        gain:
          type: number
        gainPercent:
          type: number
...